fix outdated suggestion overwriting non-suggest.
Before, a suggestion response might not be canceled when the user typed
the next character, so the outdated suggestion reached the text field
a tad too late -- and replaced the user-entered content. This became
apparent only when the next character yielded no suggestion!

Now it still overwrites but at least the non-suggestion case is handled
by displaying the .prefix case.